Add support ST_LOOP_POINT_M1
Added support for motor movement {'distance`: `xx`}

+++ cfg.cpp
@@ -53,12 +53,12 @@ uint8_t magic_number;
     }
 }
 
-uint8_t CConfig::get_distance( void )
+uint16_t CConfig::get_distance( void )
 {
     return distance;
 }
 
-void CConfig::set_distance( uint8_t val )
+void CConfig::set_distance( uint16_t val )
 {
     distance = val;
     EEPROM.put( EEPROM_ADDRESS_DISTANCE, val );
@@ -148,7 +148,7 @@ void CConfig::set_st_test( uint8_t enable )
 
 //TODO: En futuro debe admitir decimales 
 
-// {distance:'24'}       distance:0 a 254       Distancia en cm donde se aplica la fuerza.
+// {distance:'500'}       distance:0 a 254       Distancia en cm donde se aplica la fuerza.
 // {force:'11'}          force:0 a 254          Fuerza a aplicar en Kg.
 // {reaction_one:'1'}    reaction_one :0 a 254  Fuerza de reaccion uno, en Kg.
 // {reaction_two:'2'}    reaction_two :0 a 254  Fuerza de reaccion dos, en Kg.
